SmartGS_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
pydantic==2.5.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.8.0
python-multipart==0.0.6
stripe==7.8.0